+++ VCStreamCoordinator.h
@@ -13,9 +13,6 @@ NS_ASSUME_NONNULL_BEGIN
 - (CVPixelBufferRef)copyLatestPixelBuffer CF_RETURNS_RETAINED;
 - (CVPixelBufferRef)copyLatestPixelBufferWithAspectFill:(BOOL * _Nullable)aspectFill
                                            preferredFPS:(NSInteger * _Nullable)preferredFPS CF_RETURNS_RETAINED;
-- (void)publishCompatibilityOutputPixelBuffer:(CVPixelBufferRef _Nullable)pixelBuffer;
-- (CVPixelBufferRef _Nullable)copyLatestCompatibilityOutputPixelBufferWithActivePath:
-    (BOOL * _Nullable)activePath CF_RETURNS_RETAINED;
 - (BOOL)isReplacementActive;
 - (BOOL)isSystemPipelineReplacementConfigured;
 - (NSInteger)preferredFPS;